# gala-ragdoll/swagger_server/models/conf_host.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class ConfHost(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """

    def __init__(self, domain_name: str=None, host_ids: List[object]=None):  # noqa: E501
        """ConfHost - a model defined in Swagger

        :param domain_name: The domain_name of this ConfHost.  # noqa: E501
        :type domain_name: str
        :param host_ids: The host_ids of this ConfHost.  # noqa: E501
        :type host_ids: List[object]
        """
        self.swagger_types = {
            'domain_name': str,
            'host_ids': List[object]
        }

        self.attribute_map = {
            'domain_name': 'domainName',
            'host_ids': 'hostIds'
        }

        self._domain_name = domain_name
        self._host_ids = host_ids

    @classmethod
    def from_dict(cls, dikt) -> 'ConfHost':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The ConfHost of this ConfHost.  # noqa: E501
        :rtype: ConfHost
        """
        return util.deserialize_model(dikt, cls)

    @property
    def domain_name(self) -> str:
        """Gets the domain_name of this ConfHost.

        domain name  # noqa: E501

        :return: The domain_name of this ConfHost.
        :rtype: str
        """
        return self._domain_name

    @domain_name.setter
    def domain_name(self, domain_name: str):
        """Sets the domain_name of this ConfHost.

        domain name  # noqa: E501

        :param domain_name: The domain_name of this ConfHost.
        :type domain_name: str
        """

        self._domain_name = domain_name

    @property
    def host_ids(self) -> List[object]:
        """Gets the host_ids of this ConfHost.


        :return: The host_ids of this ConfHost.
        :rtype: List[object]
        """
        return self._host_ids

    @host_ids.setter
    def host_ids(self, host_ids: List[object]):
        """Sets the host_ids of this ConfHost.


        :param host_ids: The host_ids of this ConfHost.
        :type host_ids: List[object]
        """

        self._host_ids = host_ids